First implemenatation of working validation task.
This update provides some additional project structure, as well as
delivering the first version of a working 'ramlValidate' task.  The test
cases validate that a known good doc (the one from the RAML 200
tutorial) will pass validation, and  a known bad doc (one shipped with
the plugin) will cause the task to fail.